+++ packages/thekselect-vue/src/index.ts
@@ -1,2 +1,2 @@
 export { useThekSelect } from './composable';
-export { default as ThekSelect } from './ThekSelect.vue';
+export { default as ThekSelect, default } from './ThekSelect.vue';